add light class

+++ src/camera.rs
@@ -5,8 +5,8 @@ pub struct Camera {
     pub eye_position: Vec3,
     pub up_direction: Vec3,
 
-    pub film_width: f32,
-    pub film_height: f32,
+    pub film_width: f64,
+    pub film_height: f64,
     pub film_center: Vec3,
 }
 
@@ -21,7 +21,7 @@ impl Camera {
         }
     }
 
-    pub fn get_pixel_world_position(&self, x_ratio: f32, y_ratio: f32) -> Vec3 {
+    pub fn get_pixel_world_position(&self, x_ratio: f64, y_ratio: f64) -> Vec3 {
         let clamped_x_ratio = x_ratio.max(0.0).min(1.0) - 0.5;
         let clamped_y_ratio = y_ratio.max(0.0).min(1.0) - 0.5;
         let center_direction = self.film_center - self.eye_position;
@@ -32,7 +32,7 @@ impl Camera {
         return film_position
     }
 
-    pub fn generate_pixel_ray(&self, x_ratio: f32, y_ratio: f32) -> Ray {
+    pub fn generate_pixel_ray(&self, x_ratio: f64, y_ratio: f64) -> Ray {
         let mut ray = Ray::from_point_to_point(
             self.eye_position, 
             self.get_pixel_world_position(x_ratio, y_ratio)
